backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
alpaca-py==0.28.3
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
numpy==1.24.3
pandas==2.0.3
schedule==1.2.0
asyncio-mqtt==0.16.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1